requirements.txt
google-adk==1.22.1
google-genai
pydantic
pandas
python-dateutil
PyYAML
mcp
fastmcp
chromadb
requests
litellm
pandas
psycopg_pool
psycopg[binary]

